+++ Lunar/PlaceholderView.swift
@@ -9,6 +9,6 @@ import SwiftUI
 
 struct PlaceholderView: View {
     var body: some View {
-        Text("blank")
+        Text("Coming Soon")
     }
 }
