+++ Lunar/PostRowView.swift
@@ -11,6 +11,8 @@ import SwiftUI
 struct PostRowView: View {
     var post: PostElement
 
+    @State var showingPlaceholderAlert = false
+
     var body: some View {
         VStack(alignment: .leading) {
             Text(post.post.name)
@@ -37,194 +39,52 @@ struct PostRowView: View {
             }
         }
 
+        .swipeActions(edge: .trailing, allowsFullSwipe: true) {
+            Button {
+                showingPlaceholderAlert = true
+            } label: {
+                Image(systemName: "chevron.forward.circle.fill")
+            }.tint(.blue)
+        }
+
         .swipeActions(edge: .leading, allowsFullSwipe: false) {
             Button {
-                print("upvote")
+                showingPlaceholderAlert = true
             } label: {
                 Image(systemName: "arrow.up.circle")
             }.tint(.green)
             Button {
-                print("downvote")
+                showingPlaceholderAlert = true
             } label: {
                 Image(systemName: "arrow.down.circle")
             }.tint(.red)
         }
 
         .contextMenu {
-            Menu("This is a menu") {
+            Menu("Menu") {
                 Button {
-                    print("This is a menu")
+                    showingPlaceholderAlert = true
                 } label: {
-                    Text("Do something")
+                    Text("Coming Soon")
                 }
             }
 
             Button {
-                print("Something")
+                showingPlaceholderAlert = true
             } label: {
-                Text("Something")
+                Text("Coming Soon")
             }
 
             Divider()
 
             Button(role: .destructive) {
-                print("Delete")
+                showingPlaceholderAlert = true
             } label: {
                 Label("Delete", systemImage: "trash")
             }
         }
-    }
-}
-
-struct InPostCommunityHeaderView: View {
-    var community: Community
-
-    var body: some View {
-        HStack(spacing: 0) {
-            ImageViewWithPlaceholder(imageURL: community.icon, placeholderSystemName: "bookmark.square.fill")
-
-            HStack(alignment: .lastTextBaseline, spacing: 1) {
-                Text(String(community.name))
-                Text(String("@\(URLParser.extractDomain(from: community.actorID))"))
-                    .foregroundStyle(.gray).opacity(0.8)
-            }
-            .lineLimit(1)
-            .font(.subheadline)
-        }
-    }
-}
-
-struct InPostThumbnailView: View {
-    @State private var isLoading = true
-
-    let processor = DownsamplingImageProcessor(size: CGSize(width: 1300, height: 1300))
-
-    var thumbnailURL: String
-    var body: some View {
-        KFImage(URL(string: thumbnailURL))
-            .onProgress { receivedSize, totalSize in
-                if receivedSize < totalSize {
-                    isLoading = true
-                } else {
-                    isLoading = false
-                }
-            }
-            .setProcessor(processor)
-            .resizable()
-            .cancelOnDisappear(true)
-            .fade(duration: 0.2)
-            .progressViewStyle(.circular)
-            .aspectRatio(contentMode: .fit)
-            .frame(alignment: .center)
-            .clipShape(RoundedRectangle(cornerRadius: 10, style: .continuous))
-            .padding(.bottom, 3)
-    }
-}
-
-struct InPostUserView: View {
-    var text: String
-    var iconName: String
-    var userAvatar: String?
-
-    var body: some View {
-        HStack(alignment: .center, spacing: 2) {
-            ImageViewWithPlaceholder(imageURL: userAvatar, placeholderSystemName: "person.crop.square.fill")
-
-            Text(String(text))
-                .foregroundColor(.gray)
-                .textCase(.uppercase)
-        }
-        .font(.subheadline)
-        .padding(.trailing, 2)
-        .lineLimit(1)
-    }
-}
-
-struct InPostMetadataView: View {
-    var bodyText: String
-    var iconName: String
-    var iconColor: Color
-    var body: some View {
-        HStack(alignment: .firstTextBaseline, spacing: 1) {
-            Image(systemName: iconName)
-                .symbolRenderingMode(.hierarchical)
-                .foregroundColor(iconColor)
-            Text(String(bodyText))
-                .foregroundColor(iconColor)
-                .textCase(.uppercase)
-        }
-        .font(.subheadline)
-        .padding(.horizontal, 2)
-        .lineLimit(1)
-    }
-}
-
-extension String {
-    func isValidExternalImageURL() -> Bool {
-        let validURLs = [
-            "i.imgur.com",
-            "media1.giphy.com",
-            "media.giphy.com",
-            "files.catbox.moe",
-            "i.postimg.cc",
-        ]
-        for url in validURLs {
-            if contains(url) {
-                return true
-            }
-        }
-        return false
-    }
-}
-
-enum URLParser {
-    static func extractDomain(from url: String) -> String {
-        guard let urlComponents = URLComponents(string: url),
-              let host = urlComponents.host
-        else {
-            return ""
-        }
-
-        if let range = host.range(of: #"^(www\.)?([a-zA-Z0-9-]+\.[a-zA-Z]{2,})(\.[a-zA-Z]{2,})?$"#,
-                                  options: .regularExpression)
-        {
-            return String(host[range])
-        }
-
-        return ""
-    }
-}
-
-struct ImageViewWithPlaceholder: View {
-    var imageURL: String?
-    var placeholderSystemName: String
-
-    var body: some View {
-        let processor = DownsamplingImageProcessor(size: CGSize(width: 50, height: 50))
-        if let url = imageURL, let image = URL(string: url) {
-            KFImage(image)
-                .setProcessor(processor)
-                .placeholder {
-                    Image(systemName: placeholderSystemName)
-                        .resizable()
-                        .scaledToFit()
-                        .frame(width: 20, height: 20)
-                        .symbolRenderingMode(.hierarchical)
-                        .foregroundColor(.gray)
-                }
-                .resizable()
-                .frame(width: 20, height: 20)
-                .clipShape(RoundedRectangle(cornerRadius: 3, style: .continuous))
-                .scaledToFit()
-                .padding(.trailing, 5)
-        } else {
-            Image(systemName: placeholderSystemName)
-                .resizable()
-                .scaledToFit()
-                .frame(width: 20, height: 20)
-                .symbolRenderingMode(.hierarchical)
-                .foregroundColor(.gray)
-                .padding(.trailing, 5)
+        .alert("Coming soon", isPresented: $showingPlaceholderAlert) {
+            Button("OK", role: .cancel) {}
         }
     }
 }
